server/requirements.txt
Flask==1.1.1
Flask-Socketio==4.2.1
python-socketio==4.3.1